+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Ansible Changes By Release
 * updated porting docs with note about inventory_dir
 * removed extension requirement for yaml inventory plugin to restore previous behaviour
 * fixed ansible-pull to now correctly deal with inventory
+* corrected dig lookup docs
 
 <sdfasdfsadfsdflkjsdfklj3oiqrua id="2.4"></a>